+++ v1_20_1/src/main/java/me/molybdenum/ambience_mini/v1_20_1/client/core/state/CombatState.java
@@ -1,7 +1,5 @@
 package me.molybdenum.ambience_mini.v1_20_1.client.core.state;
 
-import me.molybdenum.ambience_mini.engine.client.core.setup.BaseClientConfig;
-import me.molybdenum.ambience_mini.engine.client.core.setup.ServerSetup;
 import me.molybdenum.ambience_mini.engine.client.core.state.BaseCombatState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.BossHealthOverlay;
@@ -12,6 +10,7 @@ import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.fml.util.ObfuscationReflectionHelper;
+import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.List;
 import java.util.Map;
@@ -32,6 +31,12 @@ public class CombatState extends BaseCombatState<Entity, Vec3>
     @Override
     public int getEntityId(Entity entity) {
         return 0;
+    }
+
+    @Override
+    public String getEntityResourceLocation(Entity entity) {
+        var key = ForgeRegistries.ENTITY_TYPES.getKey(entity.getType());
+        return key == null ? null : key.toString();
     }